KIWI-1696: Move files
